clean up test class organization
