+++ cosmos_service/test/src/annotations/test_sidarthe.py
@@ -3,7 +3,6 @@ import json
 
 PDF_NAME='sidarthe'
 REMOTE_URL='https://www.nature.com/articles/s41591-020-0883-7.pdf'
-AREA_BOUNDS=(0.9,1.1)
 
 class TestSidartheAnnotations(BaseAnnotationComparisonTest):
 
@@ -11,23 +10,31 @@ class TestSidartheAnnotations(BaseAnnotationComparisonTest):
         self._setup(PDF_NAME, REMOTE_URL)
 
     def test_compare_figure_counts(self):
-        fc = self.compare_pages_for_label("Figure")
-        failures = fc.get_failures_per_page(
-            fc.expected_counts, fc.cosmos_counts, equals_comparator)
-        assert failures.ok(), failures.error_message()
+        self.check_count_per_page("Figure")
 
-    def test_compare_overlapping_areas(self):
-        fc = self.compare_pages_for_label("Figure")
-        failures = fc.get_failures_per_page(
-            [AREA_BOUNDS for _ in fc.expected_counts],
-            fc.overlap_percents, 
-            in_bounds_comparator)
-        assert failures.ok(), failures.error_message()
+    def test_compare_overlapping_figure_areas(self):
+        self.check_overlap_per_page("Figure")
 
-    def test_compare_total_areas(self):
-        figure_comparisons = self.compare_pages_for_label("Figure")
-        failures = figure_comparisons.get_failures_per_page(
-            [AREA_BOUNDS for _ in figure_comparisons.expected_counts],
-            figure_comparisons.area_ratios, 
-            in_bounds_comparator)
-        assert failures.ok(), failures.error_message()
+    def test_compare_figure_caption_counts(self):
+        self.check_count_per_page("Figure Caption")
+
+    def test_compare_overlapping_figure_caption_areas(self):
+        self.check_overlap_per_page("Figure Caption")
+
+    def test_compare_table_counts(self):
+        self.check_count_per_page("Table")
+
+    def test_compare_overlapping_table_areas(self):
+        self.check_overlap_per_page("Table")
+
+    def test_compare_table_caption_counts(self):
+        self.check_count_per_page("Table Caption")
+
+    def test_compare_overlapping_table_caption_areas(self):
+        self.check_overlap_per_page("Table Caption")
+
+    def test_compare_equation_counts(self):
+        self.check_count_per_page("Equation")
+
+    def test_compare_overlapping_equation_areas(self):
+        self.check_overlap_per_page("Equation")
